Move glam-header component to `glam` directory
This component is actually what started this whole refactoring, since it
felt weird to have the glam-header and header components right next to
each other when they do not actually interact.

PER-9753: Glam Onboarding Header
