+++ src/app/onboarding/components/glam/glam-header/glam-header.component.scss
