+++ src/app/onboarding/components/glam/glam-header/glam-header.component.spec.ts
@@ -3,7 +3,7 @@ import { Shallow } from 'shallow-render';
 import { AccountService } from '@shared/services/account/account.service';
 import { Router } from '@angular/router';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { OnboardingModule } from '../../onboarding.module';
+import { OnboardingModule } from '../../../onboarding.module';
 import { GlamOnboardingHeaderComponent } from './glam-header.component';
 
 describe('GlamHeaderComponent', () => {
